minor refactoring, distractors not right
